Almost finished Found Luggage.
Getting the String value date from the excel sheet and converting it to a Date value is not working yet.
Weird bug with the String time when getting the value from an excel sheet, it makes the time 1 hour earlier than what is listed in the sheet.
There might be some work needed for the layout from the FMXL file.